Update util module
* `IncrementalMeanVarianceFromBatch`
