+++ aine_drl/util/__init__.py
@@ -1,3 +1,4 @@
 from .util_methods import *
 from .logger import *
 from .static_recursive_buffer import StaticRecursiveBuffer
+from .incremental_mean_variance_from_batch import IncrementalMeanVarianceFromBatch
